requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
torchvision==0.9.1+cpu
Flask==2.0.1
seaborn==0.11.1
tqdm==4.60.0
matplotlib==3.4.1
requests==2.22.0
numpy==1.20.2
pandas==1.2.4
Pillow==8.3.0
protobuf==3.17.3
scikit_learn==0.24.2
gunicorn
